refactor(web): consolidate shared config/constants and simplify SW wiring
Post-merge cleanup of the config and service-worker layers.

Config / constants:
- Rename config/envs.ts -> config/envs.shared.ts and expose it via a config
  barrel (config/index.ts). Make the API URL resolver work in both the window
  and the service worker (self with a window fallback).
- Remove the configuration.sw.ts / configuration.shared.ts wrappers and the
  dead SharedConfig type; the service worker now reads the envs.shared
  constants directly, making clear these are build-time constants and not the
  server-loaded config the SW does not support yet.
- Move STREAMER_URL_PREFIX into envs.shared and SW_COVERS_CACHE_KEY next to the
  cover helpers (covers/helpers.shared); delete workers/constants.shared.ts.

Service worker:
- Replace serviceWorkerReadySignal with the native
  navigator.serviceWorker.controller check (a SW can only stream for pages it
  controls); delete workers/states.web.ts.
- Extract the periodic background-task trigger into a <BackgroundTasks />
  component rendered alongside <BackgroundReplication />.

Misc:
- Relocate the Intl.DurationFormat shim into vite-env.d.ts and drop the custom
  PromiseReturnType helper in favour of Awaited<ReturnType<...>>.

tsc, biome and the full web test suite (246 tests) pass.
