--- apps/web/src/config/configuration.shared.ts
@@ -1,3 +1,0 @@
-export const getApiUrl = () =>
-  import.meta.env.VITE_API_URL ||
-  `${self.location.protocol}//${self.location.hostname}:3000`
