--- apps/web/src/config/types.shared.ts
@@ -1,4 +1,0 @@
-export type SharedConfig = {
-  API_COUCH_URI?: string
-  API_URL?: string
-}
