+++ apps/web/src/covers/useCoversCacheInformation.ts
@@ -1,6 +1,6 @@
 import { from, switchMap, map } from "rxjs"
 import { useQuery$ } from "reactjrx"
-import { serviceWorkerConfiguration } from "../config/configuration.sw"
+import { SW_COVERS_CACHE_KEY } from "./helpers.shared"
 
 export const USE_COVERS_CACHE_INFORMATION_KEY = ["storage/covers/size"]
 
@@ -10,7 +10,7 @@ export const useCoversCacheInformation = () =>
     networkMode: "always",
     gcTime: 5 * 60 * 1000,
     queryFn: () =>
-      from(caches.open(serviceWorkerConfiguration.SW_COVERS_CACHE_KEY)).pipe(
+      from(caches.open(SW_COVERS_CACHE_KEY)).pipe(
         switchMap((cache) =>
           from(cache.keys()).pipe(
             map((keys) => {
