+++ apps/web/src/covers/useRemoveCoversInCache.ts
@@ -2,14 +2,14 @@ import { useMutation$ } from "reactjrx"
 import { from, switchMap } from "rxjs"
 import { USE_COVERS_CACHE_INFORMATION_KEY } from "./useCoversCacheInformation"
 import { useQueryClient } from "@tanstack/react-query"
-import { serviceWorkerConfiguration } from "../config/configuration.sw"
+import { SW_COVERS_CACHE_KEY } from "./helpers.shared"
 
 export const useRemoveCoversInCache = () => {
   const client = useQueryClient()
 
   return useMutation$({
     mutationFn: () =>
-      from(caches.open(serviceWorkerConfiguration.SW_COVERS_CACHE_KEY)).pipe(
+      from(caches.open(SW_COVERS_CACHE_KEY)).pipe(
         switchMap((cache) =>
           from(cache.keys()).pipe(
             switchMap((keys) => {
