+++ apps/web/src/reader/manifest/getResourcePathFromUrl.shared.ts
@@ -1,4 +1,4 @@
-import { STREAMER_URL_PREFIX } from "../../workers/constants.shared"
+import { STREAMER_URL_PREFIX } from "../../config/envs.shared"
 
 const getEpubNameFromUrl = (url: string) => {
   const { pathname } = new URL(url)
