--- apps/web/src/types.ts
@@ -1,2 +1,0 @@
-export type PromiseReturnType<T extends (...args: any) => Promise<any>> =
-  T extends (...args: any) => Promise<infer U> ? U : any
