--- apps/web/src/workers/constants.shared.ts
@@ -1,1 +1,0 @@
-export const STREAMER_URL_PREFIX = `streamer`
