--- apps/web/src/workers/states.web.ts
@@ -1,5 +1,0 @@
-import { signal } from "reactjrx"
-
-export const serviceWorkerReadySignal = signal({
-  default: false,
-})
